feat: create the fetch helper and some adjustments
